% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qleTest.R
\name{checkMultRoot}
\alias{checkMultRoot}
\title{Inspect estimated parameters}
\usage{
checkMultRoot(est, par = NULL, fun = getOption("qle.fork", "lapply"),
  cl = NULL, verbose = FALSE)
}
\arguments{
\item{est}{object of class "\code{qle}", the estimation results from function \code{\link{qle}}}

\item{par}{list or matrix of roots of quasi-score vector}

\item{fun}{either "\code{lapply}" (default) or "\code{mclapply}" for parallel processing}

\item{cl}{cluster object, \code{NULL} (default), see \code{\link[parallel]{makeCluster}}}

\item{verbose}{logical, \code{TRUE} for intermediate output}
}
\value{
A data frame with columns `\code{quasi-deviance}`, `\code{Minor}`, `\code{det}`, `\code{max}`,
 `\code{trace}` and `\code{score_max}`  (see vignette). The second column shows the leading minor of
  the observed QI matrix which is not positive definite. If so, then the corresponding parameter cannot be an consistent
  estimate of the model parameter. The rows show the corresponding values for each parameter from `\code{x}`.
  Further, the results of function \code{\link{quasiDeviance}} for each parameter in `\code{x}` are
  returned as `attr(,"qdev")`. In case of any errors the corresponding (failed) results are returned as a
  named attribute `\code{X}`.
}
\description{
Check out estimated roots of the quasi-score vector and find the best one
}
\details{
In case of the quasi-likelihood approach only (using the criterion "\code{qle}")
 we can compare the observed quasi-information matrix with the actual (expected) one by (scalar equivalent)
 criteria (see the vignette). We measure the dissimilarity of both matrices evaluated at approximate
 roots `\code{x}` for which these criteria should be neglectably small if they correspond to plausible roots.
 By the smallest value of any of these criteria we extract the best root.
 Note that you can pass a weighting matrix `W` and location parameter `theta` by `\code{...}` to compute
 a variance matrix approximation, particularly, if these were part of the results of function \code{\link{qle}}.
}
\author{
M. Baaske
}
