% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qleTest.R
\name{print.qleTest}
\alias{print.qleTest}
\title{print \code{qleTest} results}
\usage{
\method{print}{qleTest}(x, pl = 1, digits = 5, ...)
}
\arguments{
\item{x}{object of class `\code{qleTest}` from call to \code{\link{qleTest}}}

\item{pl}{not used yet}

\item{digits}{number of (significant) digits to display}

\item{...}{ignored, additional arguments}
}
\description{
print the results of call to \code{\link{qleTest}}
}
