% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{QLmodel}
\alias{QLmodel}
\title{Construct quasi-likelihood model}
\usage{
QLmodel(qldata, lb, ub, obs, mods, nfit = 1, cvfit = FALSE,
  var.type = c("wcholMean", "cholMean", "wlogMean", "logMean", "kriging",
  "const"), useVar = TRUE, criterion = c("qle", "mahal"))
}
\arguments{
\item{qldata}{simulation data (see \code{\link{setQLdata}})}

\item{lb}{numeric vector of lower bounds defining the (hyper)box}

\item{ub}{numeric vector of upper bounds defining the (hyper)box}

\item{obs}{numeric vector of (real data) statistics}

\item{mods}{list of (fitted) covariance models (see \code{\link{fitSIRFk}})}

\item{nfit}{number of cycles, \code{nfit=1} (default), after which covariance
parameters are re-estimated otherwise re-used}

\item{cvfit}{logical, \code{FALSE} (default), whether to re-fit CV models (re-estimate covariance parameters)}

\item{var.type}{name of the variance approximation method (see \code{\link{covarTx}})}

\item{useVar}{logical, \code{TRUE} (default), whether to use prediction variances (see details)}

\item{criterion}{global criterion function for sampling and minimization, either "\code{qle}" or "\code{mahal}"}
}
\value{
Object of class \code{\link{QLmodel}} which stores the data frame of simulation results, bounds on
 the parameter space, covariance models for kriging, observations vector as well as options for kriging and fitting.
}
\description{
Aggregate and construct the data for quasi-likelihood estimation
}
\details{
The function aggregates all information for estimation storing the fitted
  covariance models of statistics and the type of variance matrix approximation. For advanced users
  this function explicitly offers the data structure to construct individual covariance models for
  each statistic as defined by \code{\link{setCovModel}}. The user has the choice whether or not to
  make use of of kriging prediction variances by `\code{useVar}` later for approximation
  of the variance matrix of statstics. If this is \code{TRUE}, then a kriging procedure calculating prediction 
  variances is used. Otherwise we use the so-called \emph{dual} approach which has some computational advantage when
  no prediction variances are needed.

	 See \code{\link{getQLmodel}} for an example.
}
\author{
M. Baaske
}
