% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{reml}
\alias{reml}
\title{Restricted maximum likelihood (REML)}
\usage{
reml(models, pars, data, Xs, verbose = FALSE)
}
\arguments{
\item{models}{object of class \code{krige} (list of covariance models) or
`\code{covModel}` (a single covariance model), see \code{\link{setCovModel}}}

\item{pars}{covariance parameter vector (including global scalar nugget value)}

\item{data}{data frame of simulated statistics, each column corresponds to a 
single covariance model in `\code{models}`}

\item{Xs}{matrix of sample points}

\item{verbose}{logical, if \code{TRUE}, print intermediate output}
}
\value{
List of REML function values
}
\description{
Calculate the REML function value for given covariance parameters
}
\details{
Given a list of covariance models the function calculates the REML function values at
  parameters `\code{pars}`.
}
\author{
M. Baaske
}
